INT-2612b added dedicated parser tests
